+++ src/timeseriesflattener/__init__.py
@@ -1,2 +1,9 @@
 """Init timeseriesflattener."""
+from .feature_spec_objects import (
+    OutcomeGroupSpec,
+    OutcomeSpec,
+    PredictorGroupSpec,
+    PredictorSpec,
+    TextPredictorSpec,
+)
 from .flattened_dataset import TimeseriesFlattener